VCV-183-1: inline hooks into review export list
